pyproject.toml
[tool.poetry]
name = " coursework"
version = "0.1.0"
description = ""
authors = ["Andrey Maishev <fatym98@gmial.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
logging-config = "^1.0.4"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pytest = "^8.3.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.14.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
